Reorganize superpowers patches
So far we had patches only for the app part of Superpowers, but there
will be patches for the other parts coming (for example for the core
part). So to avoid confusion which patches should be applied to which
part of Superpowers, put the existing patches for the app part in the
app-patches subdirectory.